refactor: move getTimeToLive helper function to the utils folder
