+++ src/mdlDocumentBuilder/helpers/defaultDates.ts
@@ -48,7 +48,3 @@ export function getDateParts(date: Date): DateParts {
     year: date.getFullYear().toString(),
   };
 }
-
-export function getTimeToLiveEpoch(ttlMinutes: number): number {
-  return Math.floor((Date.now() + ttlMinutes * 60 * 1000) / 1000);
-}
