+++ src/utils/getTimeToLiveEpoch.ts
@@ -1,0 +1,3 @@
+export function getTimeToLiveEpoch(ttlMinutes: number): number {
+  return Math.floor((Date.now() + ttlMinutes * 60 * 1000) / 1000);
+}
